refactor(dashboard page_run.py) split into modules
splitting the code improves maintainability

also improve loggign
