+++ odtp/dashboard/main.py
@@ -6,7 +6,7 @@ from odtp.dashboard.page_about.main import content as about_page
 from odtp.dashboard.page_components.main import content as components_page
 from odtp.dashboard.page_digital_twins.main import content as digital_twins_page
 from odtp.dashboard.page_executions.main import content as executions_page
-from odtp.dashboard.pages.page_run import content as run_page
+from odtp.dashboard.page_run.main import content as run_page
 from odtp.dashboard.page_users.main import content as user_page
